extract addNewParamToSearchPath from SearchView
