+++ projects/bp-gallery/environments/.env.development
@@ -1,3 +1,3 @@
-REACT_APP_API_BASE=https://bilder-dev.bad-harzburg-stiftung.de/api
+REACT_APP_API_BASE=https://bilder.bad-harzburg-stiftung.de/api
 REACT_APP_CONTACT_MAIL=archiv@bad-harzburg-stiftung.de
 REACT_APP_ADVANCED_SEARCH=false
